requirements.txt
python-telegram-bot[asyncio]==20.7
python-dotenv==1.0.0
aiosqlite==0.19.0